pyproject.toml
[project]
name = "pygame-voxel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "moderngl>=5.12.0",
    "numba>=0.61.2",
    "pygame-ce>=2.5.3",
    "pyglm>=2.8.2",
]
